style fixes + bailey input
